Files restructure (#1187)
