+++ src/shared/configuration/internal.ts
