+++ src/shared/configuration/jsonConfig.ts
