+++ src/shared/logging/internalAzureLogger.ts
