+++ src/shared/util/fileSystemHelper.ts
